refactor(format): abstract formatting conditions
